requirements.txt
uWSGI
flask
flask_sqlalchemy
flask_marshmallow
SQLAlchemy
PyJWT
flask_jwt_extended
uuid
werkzeug
flask-restx
Flask-WTF
wtforms
marshmallow
Flask-CORS
boto3
Pillow
python-dotenv
faker
pymysql
marshmallow-sqlalchemy
Flask-Migrate